+++ src/chai/ManagedArray.hpp
@@ -15,6 +15,7 @@
 
 #include "umpire/Allocator.hpp"
 
+#include <concepts>
 #include <cstddef>
 
 namespace chai
@@ -76,6 +77,13 @@ template <typename T>
 class ManagedArray : public CHAICopyable
 {
 public:
+  using size_type = size_t;
+  using value_type = T;
+  using reference = T&;
+  using const_reference = const T&;
+  using pointer = T*;
+  using const_pointer = const T*;
+
   using T_non_const = typename std::remove_const<T>::type;
 
   CHAI_HOST_DEVICE ManagedArray();
@@ -98,7 +106,9 @@ public:
    * \param elems Number of elements in the array.
    * \param space Execution space in which to allocate the array.
    */
-  CHAI_HOST_DEVICE ManagedArray(size_t elems, ExecutionSpace space = get_default_space());
+  CHAI_HOST_DEVICE explicit ManagedArray(
+      size_t elems,
+      ExecutionSpace space = get_default_space());
 
   ManagedArray(
       size_t elems,
@@ -126,8 +136,17 @@ public:
 
   /*!
    * \brief Construct a ManagedArray from a nullptr.
+   *
+   * \note The constraint prevents overload ambiguity between this
+   *       constructor and the size_t constructor when constructing
+   *       with the integer literal 0. Additionally, the constraint
+   *       is a workaround for a nvcc bug where overload resolution
+   *       can incorrectly report an ambiguity between this constructor
+   *       and the size_t constructor.
+   *
+   * \todo Consider removing this constructor.
    */
-  CHAI_HOST_DEVICE ManagedArray(std::nullptr_t other);
+  CHAI_HOST_DEVICE ManagedArray(std::same_as<std::nullptr_t> auto) : ManagedArray() {}
 
   CHAI_HOST ManagedArray(PointerRecord* record, ExecutionSpace space);
 
@@ -380,7 +399,7 @@ public:
   /*!
   * Accessor for m_is_slice -whether this array was created with a slice() command.
   */
-  CHAI_HOST_DEVICE bool isSlice() { return m_is_slice;}
+  CHAI_HOST_DEVICE bool isSlice() const { return m_is_slice;}
 
 
 private: